requirements.txt
dill==0.2.9
numpy==1.17.0
joblib==0.13.2
matplotlib==3.0.3
natsort==6.0.0
pandas==0.24.2
psycopg2==2.8.3
pulp==2.0
rpy2==3.0.5
scikit-image==0.14.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.10.0
statsmodels==0.10.0rc2
stopit==1.1.2
elitech-datareader==1.0.4
